Refactor transformer, update export conditions
